UAV-Mapping-README
UAV-Mapping-README
===========

Background
-----------------------------------------------------------------------------------------------------------------
UAV mapping will be used to assist in the systems Artificial Intelligence 
by applying a combination of Labview's vision and motion module functions with wireless real 
time UAV camera feeds.
This will track and process the robots as well as any dynamic threats in the system. 
The data will be acquired from an UAV with attached wireless webcam,
processed at the control centre with relevant images tracked using vision assistant and plotted 
using an XY graph to simulate a birds eye view of the map.
This will be used in conduction with other intelligence gathered on the ground to complement automation. 

Idea Generation Tools
------------------------------------------------------------------------------------------------------------------
Idea was generated and validated as a group within the initial design stages.
The tools to generate this idea primary consisted of applying a divergent thinking approach based
around fixing the integration gaps acquired in N2 analysis and relating the system to a battlefield which
highlighted the similarities and differences resulting in the idea of creating a form aerial visual processing.
Potential issues such as sufficient altitude where tested before major built commenced.

Guide to the Repository
------------------------------------------------------------------------------------------------------------------
Labview code for the AI Mapping system will be uploaded to here.
Relavent Systems Architecture specific to this subsystem will be located here also